drivers: timer: nrf_rtc_timer: Apply workaround for Errata 165 (WDT)
Workaround is applied for nrf53net. It sets a PPI connection
between RTC event and WDT START task. An event must occur one tick
before RTC event that results in an interrupt. In case of RTC this
can be CC or OVERFLOW. Current solution is using dedicated channel
for CC0 and second dedicated channel for OVERFLOW. It means that
user channels API cannot be used when workaround is applied.

Workaround assumes that WDT is not used on nrf53 cpunet.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>